+++ src/Untappd.Net/Authentication/IAuthenticatedUntappdCredentials.cs
@@ -1,4 +1,4 @@
-﻿namespace Untappd.Net.Client
+﻿namespace Untappd.Net.Authentication
 {
     public interface IAuthenticatedUntappdCredentials : IUntappdCredentials
     {
