+++ src/Untappd.Net/Authentication/IUnAuthenticatedUntappdCredentials.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace Untappd.Net.Client
+namespace Untappd.Net.Authentication
 {
     public interface IUnAuthenticatedUntappdCredentials : IUntappdCredentials
     {
