+++ src/Untappd.Net/Authentication/UntappdCredentials.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
+﻿using System.Collections.Generic;
 
-namespace Untappd.Net.Client
+namespace Untappd.Net.Authentication
 {
     public abstract class UntappdCredentials : IUntappdCredentials
     {
